Exception Handler added

+++ src/main/java/com/illutech/advertsite/controller/UsersController.java
@@ -1,6 +1,7 @@
 package com.illutech.advertsite.controller;
 
 import com.illutech.advertsite.entities.Users;
+import com.illutech.advertsite.error.UserNotFoundException;
 import com.illutech.advertsite.models.UsersModel;
 import com.illutech.advertsite.service.userservice.IUsersService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -47,8 +48,10 @@ public class UsersController {
         return tmp;
     }
     @PostMapping("/DeleteById/{id}")
-    public Users DeleteById(@PathVariable Long id){
-        return usersService.DeleteById(id);
+    public Users DeleteById(@PathVariable Long id) throws UserNotFoundException {
+        if (usersService.exists(id)) return usersService.DeleteById(id);
+        else throw new UserNotFoundException("کاربر مورد نظر پیدا نشد.");
+
     }
     @GetMapping("/getUsers")
     public Collection<Users> getUsersList(){
